Fix writing of 64-bit variables to trace.

+++ write-trace.c
@@ -95,7 +95,7 @@ void write_trace_variables(FILE *out, uint32_t n_vars, ...)
               }
           case 8:
               {
-                  int64_t val = va_arg(ap, int);
+                  int64_t val = va_arg(ap, int64_t);
                   fwrite(&val, sizeof(val), 1, out);
                   break;
               }
